Fix and enable spice cloud tests
